\name{rstdnorm}
\alias{rstdnorm}
\title{
Random Sample from the Standard Normal Distribution
}
\description{
This function returns a single draw from the standard normal distribution.
}
\usage{
rstdnorm()
}
\value{
A numeric vector of length one.
}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
\examples{
rstdnorm()
}

